Fix custom_ease_system does not correctly remove EasingComponent (#55)
The call to `remove()` in `custom_ease_system` uses the wrong type
variable in `EasingComponent` so that the removal has no effect. Hence,
the easing never stops processing and continues mark the target
component as mutated every frame. This causes additional work for any
other systems which are watching for changes to that component.

Co-authored-by: François Mockers <francois.mockers@vleue.com>